feat(kre-go): refactor for better scope of context methods.
